Refactors vehicle registration logic and cleans up SQL queries
Enhances the vehicle registration process by integrating the Banco class directly into the main method and improving input validation throughout the CarroDao and MotocicletaDao classes.

Adjusts SQL query formatting for consistency and correctness, ensuring proper data types are used for vehicle attributes.

Removes unnecessary status validation code for clarity and streamlines the user experience in the EstoqueView for vehicle management options.

+++ src/Models/Carro.java
@@ -1,24 +1,24 @@
 package models;
 
 public class Carro extends Veiculo {
-    private double cavaloPotencia;
+    private int cavaloPotencia;
     private int numeroPortas;
     private int idTipoCombustivel;
     private String[] tipoCombustivel = {"Gasolina", "Etanol", "Diesel", "Elétrico"};
 
     public Carro(String modelo, String numChassi, double quilometragem, double preco, String cor, 
-                int anoFabricacao, int idStatus, double cavaloPotencia, int numeroPortas, int idTipoCombustivel) {
+                int anoFabricacao, int idStatus, int cavaloPotencia, int numeroPortas, int idTipoCombustivel) {
         super(modelo, numChassi, quilometragem, preco, cor, anoFabricacao, idStatus);
         this.cavaloPotencia = cavaloPotencia;
         this.numeroPortas = numeroPortas;
         this.idTipoCombustivel = idTipoCombustivel;
     }
 
-    public double getCavaloPotencia() {
+    public int getCavaloPotencia() {
         return cavaloPotencia;
     }
 
-    public void setCavaloPotencia(double cavaloPotencia) {
+    public void setCavaloPotencia(int cavaloPotencia) {
         this.cavaloPotencia = cavaloPotencia;
     }
